v1
git-svn-id: svn://orca.physics.unc.edu/Orca@9517 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/VMEcpu/VME_Readout_Code/ORCAEN830Readout.cc
@@ -15,8 +15,8 @@ bool ORCAEN830Readout::Start()
 
 bool ORCAEN830Readout::Readout(SBC_LAM_Data* lamData)
 {
-    uint32_t statusRegOffset		= GetDeviceSpecificData()[1];
-    uint32_t chan0Offset		    = GetDeviceSpecificData()[5];
+    uint32_t statusRegOffset	= GetDeviceSpecificData()[1];
+    int32_t chan0Offset		    = (int32_t)GetDeviceSpecificData()[5];
     uint16_t statusWord;
 
     int32_t result = VMERead(GetBaseAddress() + statusRegOffset,0x39, sizeof(statusWord),statusWord);
